Fix for SelectedItem initial shift
After loading initially, if the order loaded was not the first entry, then the ViewModel would load that first order before switching back to the previous order, due to the UI initially loading the ComboBox and setting the index to 0, before updating from the binding and swapping back to 2 (or whatever it was before). Binding to SelectedValue instead seems to avoid that issue.